pyproject.toml
[project]
name = "kdpo-exploration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "causal-conv1d>=1.6.1",
    "datasets>=4.7.0",
    "deepspeed>=0.18.7",
    "flash-linear-attention>=0.4.2",
    "immutabledict>=4.3.1",
    "langdetect>=1.0.9",
    "nltk>=3.9.3",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "torch",
    "transformers>=5.3.0",
    "wandb>=0.25.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130" },
]
causal-conv1d = { url = "https://github.com/Dao-AILab/causal-conv1d/releases/download/v1.6.1.post4/causal_conv1d-1.6.1+cu13torch2.10cxx11abiTRUE-cp312-cp312-linux_x86_64.whl" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
